fix trailing whitespace, rename package name, update misc files
