+++ .gitignore
@@ -1,4 +1,3 @@
 # ENV
-*.env
-
+config/*.yaml
