+++ .pre-commit-config.yaml
@@ -7,7 +7,7 @@ repos:
       - id: check-yaml
       - id: check-added-large-files
   - repo: https://github.com/golangci/golangci-lint
-    rev: v1.50.1
+    rev: v1.53.3
     hooks:
       - id: golangci-lint
         name: golangci-lint
