+++ internal/repositories/user/user_repository.go
@@ -1,4 +1,4 @@
-package user_repository
+package user
 
 import (
 	"gorm.io/gorm"
