+++ internal/services/services.go
@@ -9,27 +9,12 @@ type Services struct {
 	UserService us.UserService
 }
 
-// type ServicesConfig struct {
-// 	UserService us.UserService
-// }
-
-// func New(hc *ServicesConfig) *Services {
-// 	return &Services{
-// 		UserService: hc.UserService,
-// 	}
-// }
-
 func InitServices(r *repositories.Repositories) *Services {
-
 	userService := us.NewUserService(
 		&us.UserServiceConfig{
 			UserRepository: r.UserRepository,
 		},
 	)
-
-	// s := New(&ServicesConfig{
-	// 	UserService: userService,
-	// })
 
 	s := &Services{
 		UserService: userService,
