+++ internal/services/user/get_user.go
@@ -1,4 +1,4 @@
-package user_service
+package user
 
 func (u *userService) GetUser() string {
 	user := u.userRepository.GetUser()
